requirements.txt
Appium-Python-Client==0.44
pytest==4.6.3
selenium==	3.14.0